requirements.txt
Flask==3.1.2
numpy==2.2.6
opencv-python-headless
pillow==12.1.0
gunicorn==21.2.0
